requirements.txt
requests
numpy
pandas
python-telegram-bot
python-dotenv
